Added PokemonAnimation and updated various classes

+++ PMDMMOc/src/com/darkxell/client/renderers/floor/DungeonItemsRenderer.java
@@ -28,7 +28,7 @@ public class DungeonItemsRenderer extends AbstractRenderer
 	@Override
 	public void render(Graphics2D g, int width, int height)
 	{
-		int xStart = this.x() / TILE_SIZE, yStart = this.y() / TILE_SIZE;
+		int xStart = (int) (this.x() / TILE_SIZE), yStart = (int) (this.y() / TILE_SIZE);
 
 		for (int x = xStart; x <= xStart + width / TILE_SIZE + 1; ++x)
 			for (int y = yStart; y <= yStart + height / TILE_SIZE + 1; ++y)
